# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Station {
  id: string;
  name: string;
  lat: number;
  lng: number;
  aqi: number;
}

export interface PollutantReading {
  PM25: number;
  PM10: number;
  NO2: number;
  SO2: number;
  CO: number;
  O3: number;
}

export interface SourceAttribution {
  source: string;
  pct: number;
}

export interface ForecastPoint {
  hour: number;
  aqi: number;
  lower: number;
  upper: number;
}

export type HotspotSeverity = typeof HotspotSeverity[keyof typeof HotspotSeverity];


export const HotspotSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
} as const;

export interface Hotspot {
  zone: string;
  sourceType: string;
  confidence: number;
  lat: number;
  lng: number;
  severity: HotspotSeverity;
}

export interface WeatherData {
  windSpeed: number;
  windDir: string;
  temp: number;
  humidity: number;
}

export interface CityData {
  id: string;
  name: string;
  lat: number;
  lng: number;
  currentAQI: number;
  dominantPollutant: string;
  pollutants: PollutantReading;
  stations: Station[];
  sourceAttribution: SourceAttribution[];
  forecast: ForecastPoint[];
  hotspots: Hotspot[];
  weather: WeatherData;
  trend30d: number[];
}

export interface AdvisoryInput {
  cityName: string;
  aqi: number;
  pollutants: PollutantReading;
}

export interface AdvisoryResult {
  english: string;
  hindi: string;
}

export interface NoticeInput {
  zone: string;
  sourceType: string;
  cityAQI: number;
  confidence: number;
  cityName: string;
}

export interface NoticeResult {
  notice: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  messages: ChatMessage[];
  cityData: CityData;
}

export interface ChatResult {
  reply: string;
}