# lib/api-zod/src/generated/types/advisoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PollutantReading } from './pollutantReading';

export interface AdvisoryInput {
  cityName: string;
  aqi: number;
  pollutants: PollutantReading;
}