# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { CityData } from './cityData';

export interface ChatInput {
  messages: ChatMessage[];
  cityData: CityData;
}