# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}