# lib/api-zod/src/generated/types/cityData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ForecastPoint } from './forecastPoint';
import type { Hotspot } from './hotspot';
import type { PollutantReading } from './pollutantReading';
import type { SourceAttribution } from './sourceAttribution';
import type { Station } from './station';
import type { WeatherData } from './weatherData';

export interface CityData {
  id: string;
  name: string;
  lat: number;
  lng: number;
  currentAQI: number;
  dominantPollutant: string;
  pollutants: PollutantReading;
  stations: Station[];
  sourceAttribution: SourceAttribution[];
  forecast: ForecastPoint[];
  hotspots: Hotspot[];
  weather: WeatherData;
  trend30d: number[];
}