# lib/api-zod/src/generated/types/forecastPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ForecastPoint {
  hour: number;
  aqi: number;
  lower: number;
  upper: number;
}