# lib/api-zod/src/generated/types/hotspot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { HotspotSeverity } from './hotspotSeverity';

export interface Hotspot {
  zone: string;
  sourceType: string;
  confidence: number;
  lat: number;
  lng: number;
  severity: HotspotSeverity;
}