# lib/api-zod/src/generated/types/station.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VayuSense Urban Air Quality Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Station {
  id: string;
  name: string;
  lat: number;
  lng: number;
  aqi: number;
}